% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BISSpkg.R
\name{BI_SS}
\alias{BI_SS}
\title{Bootstrap Imputation and Variable Selection}
\usage{
BI_SS(X, Y, family, link = NULL, missing_col, MI.method, BI.size, pi, nsteps)
}
\arguments{
\item{X}{a n x p matrix of n observation and p predictors, where NAs present as missing values}

\item{Y}{a n x 1 vector represent the outcome parameter, esponse variable. Quantitative for family="gaussian", or family="poisson" (non-negative counts).
For family="binomial" should be either a factor with two levels, or a two-column matrix of counts or proportions (the second column is treated as the target class; 
for a factor, the last level in alphabetical order is the target class). 
For  "binomial" case, if y is presented as a vector, it will be coerced into a factor.}

\item{family}{of regression, could be one of the following values: "gaussian","binomial","poisson")}

\item{missing_col}{index of the missing column in X (only one column is missing in this case)}

\item{MI.method}{the type of imputation method.  Can be "mice", or  "balasso".}

\item{BI.size}{number of bootstrap imputated datasets}

\item{pi}{threshold value for adaptive lasso variable selection}

\item{nsteps}{number of steps used in stability selection}

\item{reg.type}{the type of regression used in the second step, could be lasso, lasso.adp}
}
\value{
a list of output values including Bootstrap imputated dataset(Bimp), Final variable selected indicator(S.fin), 
        final parameter estimates for the regression(beta.fin), final selection probability for lasso (S1.prob), 
        final selection for variable selection indicator for stability selection with randomized lasso(S2.prob)
}
\description{
This function deploys bootstrap imputation and variable selection on (X,y) data with missing values
}
\details{
This function first performs the bootstrap imputation based on mice method(mice package) and balasso method(MIHD) method on data consists of  
outcome parameter and predictors with missing values to obtain bootstrap imputed dataset.  Then this function use lasso and stability selection with randomized lasso to conduct
variable selection on the bootstrap imputed datasets, and obtain selection indicators. Lastly, this function calculate selected predictor estimates.
}
\examples{
 output_lr_1=BI_SS(lr_1[,-1],lr_1$Y,family="gaussian",link=NULL,missing_col=1,MI.method="mice",BI.size=100,pi=NULL,nsteps=12)
}
\author{
Zhifan Sang
}
\seealso{
\code{MIdurr}
}

