% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BISSpkg.R
\name{detect.missing}
\alias{detect.missing}
\title{Finding the indicator matrix for missing values}
\usage{
detect.missing(x)
}
\arguments{
\item{x}{a n x p matrix of n observation and p predictors, where NAs present as missing values}
}
\value{
a n x p binary indicator matrix where 1 indicates value is missing, 0 indicates value is not missing.
}
\description{
This function develops a simple way to generate the indicator matrix for matrix with missing values
}
\details{
This function runs NA search and replace on the input matrix, return a binary matrix of the same size.  If the (i,j) element 
in the input matrix is NA, then the (i,j) element of output matrix is 1; Otherwise, the (i,j) element of the output matrix is 0
}
\author{
Zhifan Sang
}
\seealso{
\code{is.na}
}

